Reorganized some stuff
