+++ __tests__/transformers/simplify_node.test.js
@@ -1,5 +1,5 @@
 describe(`simplify_node.js`, () => {
-  const simplify_node = require('../../app/transformers/simplify_node.js');
+  const simplify_node = require('../../src/transformers/simplify_node.js');
 
   it(`is a function`, () => {
     expect(typeof(simplify_node)).toEqual('function');
