+++ __tests__/utils/utils.test.js
@@ -1,4 +1,4 @@
-const { makeNode } = require('../../test-utils/utils.js');
+const { makeNode } = require('../../utils/utils.js');
 
 describe(`test-utils.js`, () => {
   describe(`makeNode`, () => {
