--- app/transformers/simplify_node.js
@@ -1,29 +1,0 @@
-const TerminalNodeImpl = require('antlr4/tree/Tree.js').TerminalNodeImpl;
-
-module.exports = function(lang_runtime_config, root) {
-
-    // Takes an antlr node generated by the antlr parser, and
-    // outputs our simplified node.
-    const simplify_node = function(node) {
-        if (node instanceof TerminalNodeImpl) {
-            return {
-                'type': lang_runtime_config.symbol_name_map[node.symbol.type + 2],
-                'begin': node.symbol.start,
-                'end': node.symbol.stop + 1,
-                'tags': [],
-                'children': [],
-                'text': node.symbol.text,
-            };
-        } else {
-            return {
-                'type': lang_runtime_config.rule_name_map[node.ruleIndex],
-                'begin': node.start.start,
-                'end': (node.stop ? node.stop : node.start).stop + 1,
-                'tags': [],
-                'children': node.children ? node.children.map(simplify_node) : [],
-            };
-        }
-    };
-
-    return simplify_node(root);
-};
